% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IEDB_CIImm.R
\name{CIImm}
\alias{CIImm}
\title{CIImm 
Implementation of the immunogenicity score from
Calis, Jorg J A et al. “Properties of MHC class I presented peptides that enhance immunogenicity.” 
PLoS computational biology vol. 9,10 (2013): e1003266. doi:10.1371/journal.pcbi.1003266
Up to now only 9-mer peptides}
\usage{
CIImm(pep)
}
\arguments{
\item{pep}{string of 9-mer peptide. ie. "SLFNTVATL"}
}
\value{
the immunogenic score
}
\description{
CIImm 
Implementation of the immunogenicity score from
Calis, Jorg J A et al. “Properties of MHC class I presented peptides that enhance immunogenicity.” 
PLoS computational biology vol. 9,10 (2013): e1003266. doi:10.1371/journal.pcbi.1003266
Up to now only 9-mer peptides
}
