% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.OpenConfigFile}
\alias{.OpenConfigFile}
\title{internal function to load configuration info.}
\usage{
.OpenConfigFile(cf)
}
\value{
the configuration list. Each slot holds one software
}
\description{
internal function to load configuration info.
}
\details{
It will create a ITSNdb.RDS file with software location paths and version
 PLEASE DO NOT DELET this file
}
