% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMHCpan.R
\name{install_netMHCPan}
\alias{install_netMHCPan}
\title{install_netMHCPan}
\usage{
install_netMHCPan(file = NULL, dir = "./")
}
\arguments{
\item{file}{= NULL}

\item{dir}{= "./"}
}
\description{
install_netMHCPan
}
