% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixMHCpred.R
\name{Install_PRIME}
\alias{Install_PRIME}
\title{Install_PRIME v 2.2
This function will install PRIME (PRedictor of class I IMmunogenic Epitopes) in your local computer. Up to now only for linux environment.
\code{\link[PRIME web site]{https://github.com/GfellerLab/PRIME}}}
\usage{
\dontrun{
library(ITSNdb)
dir.path <- "my/directory"
Install_PRIME(dir = dir.path)
}
}
\arguments{
\item{dir}{= "./", the path to the directory where do you whant to store your netMHCpan. Then it will be 
accesible blindly}
}
\value{
on success the following msg will be printed "PRIME has been successfully installed"
}
\description{
Install_PRIME v 2.2
This function will install PRIME (PRedictor of class I IMmunogenic Epitopes) in your local computer. Up to now only for linux environment.
\code{\link[PRIME web site]{https://github.com/GfellerLab/PRIME}}
}
