% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITSNdb.R
\docType{data}
\name{ITSNdb}
\alias{ITSNdb}
\title{ITSNdb}
\format{
A data frame with 199 rows (neoantigens) and 11 variables (neoantigens info)
}
\usage{
load(ITSNdb)
}
\description{
The New Immunogenic Tumor Specific Neoantigens database
holding only neoantigens whose MCH-I binding was experimentally validated and resulted positive, with immunogenicity assay performed
}
\details{
\itemize{
\item Tumor: Tumor where the neoantigen was reported
\item Author: Name of the first author in reference manuscript
\item Paper: URL of the reference manuscript
\item Neoantigen: Aminoacid sequence of the mutated neoantigen
\item WT: Aminoacid sequence of the wild type peptide originating the neoantigen
\item HLA: HLA associated to the neoantigen
\item NeoType: Positive if the neoantigen immunogenicity assay resulted positive, Negative otherwise
\item mutPosition: Position of the mutation in the neoantigen aminoacid sequence
\item PositionType: Position type of the mutation, Anchor position if MutPosition = 2 or 9, non-Anchor otherwise
\item GeneSymbol: Symbol of the gene that codes for the protein of origin
\item Length: Neoantigen length
}
}
\keyword{datasets}
