% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITSNdb.R
\docType{data}
\name{MHC_allele_names}
\alias{MHC_allele_names}
\title{MHC_allele_names.txt}
\format{
A data frame with 1 column and 3584 rows
}
\usage{
load(MHC_allele_names)
}
\description{
the New Immunogenic Tumor Specific Neoantigens database
Only neoantigens whose MCH-I binding was experimentally validated and resulted positive and they also have the immunologic assay performed
}
\details{
\itemize{
\item Alleles : all the possible alleles names for MHC-I
}
}
\keyword{datasets}
