% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMHCpan.R
\name{RunNetMHCPan_peptides}
\alias{RunNetMHCPan_peptides}
\title{RunNetMHCPan_peptides}
\usage{
RunNetMHCPan_peptides(
  peps,
  alleles,
  rthParam = 0.5,
  rltParam = 2,
  tParam = -99.9002,
  nCores = 1L
)
}
\arguments{
\item{peps}{vector of peptide sequences}

\item{alleles}{a character vector with the HLA sequences  c("HLA-A01:01") or c("HLA-A01:01","HLA-A02:01") and so on. If NULL it will be automatically downloaded from the netMHCpan server (it may take time)}

\item{rthParam}{float (default 0.5) upper limit for strong binder (SB -> peptide percent rank < rhtParam)}

\item{tParam}{tparam}

\item{nCores}{}

\item{rlhParam}{float (default 2.0) upper limit for weak binder (WB ->  rhtParam <= peptide percent rank < rltParam )}
}
\value{
a named list with data frames slots holding the netMHCpan 4.1 results table. 
Each slot corresponds to each allele and is named accordingly
length(list) == length(alleles)
}
\description{
Run netMHCpan for all the peptides against all the HLA alleles
It will provide a list holding the results for each allele agains all the peptides
The lists is named by allele
now running for netMHCpan 4.1
}
\details{
run netMHCpan trhough the sequence fasta file for each HLA
}
