% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMHCpan_cohorts.R
\name{RunNetMHCPan_pMHC}
\alias{RunNetMHCPan_pMHC}
\title{RunNetMHCPan_pMHC}
\usage{
RunNetMHCPan_pMHC(
  pMHCfile,
  rthParam = 0.5,
  rltParam = 2,
  tParam = -99.9002,
  nCores = 1L,
  outFile
)
}
\arguments{
\item{pMHCfile}{character with the file path of the fasta file (it should be .fasta and contain onle 1 sequence)}

\item{rthParam}{float (default 0.5) upper limit for strong binder (SB -> peptide percent rank < rhtParam)}

\item{tParam}{tparam}

\item{outFile}{file name of the output file}

\item{rlhParam}{float (default 2.0) upper limit for weak binder (WB ->  rhtParam <= peptide percent rank < rltParam )}
}
\value{
a character
}
\description{
Run a predict binding affinity and rank for each peptide-MHC allele pair in a cohort based manner present in the file
}
\details{
run netMHCpan trhough the sequence fasta file for each HLA
}
