% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixMHCpred.R
\name{RunPRIME}
\alias{RunPRIME}
\title{RunPRIME
It will runs PRIME immunogenic predictions for peptides-HLA pairs stored in a file
The file (a text comma separated file) should have the following columns
Sample,Neoantigen,HLA
Sample: Identify the sample, patients or any other useful annotation
Neoantigen: the neoantigen sequence (8 to 14 mer length)
HLA : the specif allele for such neoantigen
Any further column is allowed and will be kept on the final result data frame}
\usage{
RunPRIME(pepFile)
}
\arguments{
\item{pepFile}{the full path to the peptide-HLA pairs}
}
\description{
RunPRIME
It will runs PRIME immunogenic predictions for peptides-HLA pairs stored in a file
The file (a text comma separated file) should have the following columns
Sample,Neoantigen,HLA
Sample: Identify the sample, patients or any other useful annotation
Neoantigen: the neoantigen sequence (8 to 14 mer length)
HLA : the specif allele for such neoantigen
Any further column is allowed and will be kept on the final result data frame
}
