% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMHCpan_cohorts.R
\name{RunNetMHCPan_pMHC}
\alias{RunNetMHCPan_pMHC}
\title{RunNetMHCPan_pMHC}
\usage{
RunNetMHCPan_pMHC(pMHCfile, nCores = 1L, outFile = FALSE)
}
\arguments{
\item{pMHCfile}{character with the file path of csv file containing all the Neoantigen-HLA pairs and the Sample identification}

\item{nCores}{integer indicating the number of CPUs.}

\item{outFile}{boolean default TRUE. If TRUE, the same input file will be returned by adding the appropriate estimated columns}
}
\value{
a data frame with all the peptide information from the pMHCfile with the extra columns returned by netMHCpan v.1
if outFile == TRUE, it will overwrite the input file with the extra columns returned by netMHCpan 4.1
}
\description{
The csv file (comma separated) should have the following columns :
Sample,Neoantigen,HLA
Pat1,GRIAFFLKY,HLA-B27:05
...    
This columns are mandatory and the file could include others
The alleles will be checked. If they are not present in the allele Database they will be omitted
}
\details{
Run netMHCpan 4.1 to predict binding affinity and rank of each peptide-MHC allele pair in a cohort based manner present in the file.
}
